LayoutTests/media/video-load-networkState.html

<video controls></video>
<script src=video-test.js></script>
<script>
    function testNetworkState(expected, endIt)
    {
        testExpected("video.networkState", expected);
        if (endIt)
            endTest();
    }

    testExpected("video.networkState", HTMLMediaElement.EMPTY);
    waitForEvent("loadstart", function () { testNetworkState(HTMLMediaElement.LOADING); } );
    waitForEvent("loadedmetadata", function () { testNetworkState(HTMLMediaElement.LOADED_METADATA); } );
    waitForEvent("loadedfirstframe", function () { testNetworkState(HTMLMediaElement.LOADED_FIRST_FRAME); } );
    waitForEvent("load", function () { testNetworkState(HTMLMediaElement.LOADED, true); } );
    video.src = "content/test.mp4";
</script>
